requirements.txt
openai
pandas
Pillow
beautifulsoup4==4.12.3

